Assignment 5/requirements.txt
pandas
numpy
transformers
tensorflow
sentence-transformers
datasets
scikit-learn
PyYAML